+++ FauxmoESP/CHANGELOG.md
@@ -3,6 +3,36 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.3.0] 2017-11-08
+### Fixed
+- Answer correctly to GetBinaryState queries (thanks to John McC)
+
+### Deprecated
+- Use onSetState callback instead of onMessage callback
+- Use onGetState callback instead of setState method
+
+## [2.2.1] 2017-10-25
+### Fixed
+- Only change state if request is a SetBinaryState action
+
+## [2.2.0] 2017-09-01
+### Fixed
+- Remove dependency on ESPAsyncWebServer
+
+### Added
+- Option to rename devices
+
+## [2.1.1] 2017-08-25
+### Fixed
+- Call UDP beginMulticast on onStationModeGotIP
+
+## [2.1.0] 2017-02-15
+### Added
+- ESPAsyncWebServer to manage TCP connections
+
+### Fixed
+- Memory leaks in UDP and TCP connections
+
 ## [2.0.0] 2017-01-05
 ### Added
 - Different discovery strategies to maximize device discovery